vm_method.c: suppress warning
* vm_method.c (rb_frame_visibility_check): suppress comparison
  warning by gcc 4.8 and 4.9.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@50706 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ vm_method.c
@@ -917,7 +917,7 @@ rb_frame_visibility_test(rb_method_flag_t flag)
 static int
 rb_frame_visibility_check(rb_method_flag_t flag)
 {
-    return CREF_VISI(rb_vm_cref()) == flag;
+    return CREF_VISI(rb_vm_cref()) == (long)flag;
 }
 
 void
